Remove the SDK dep from otelmux testing
