+++ instrumentation/github.com/gorilla/mux/otelmux/mux_test.go
@@ -17,7 +17,6 @@ package otelmux
 import (
 	"bufio"
 	"context"
-	"fmt"
 	"io"
 	"net"
 	"net/http"
@@ -26,12 +25,9 @@ import (
 
 	"github.com/gorilla/mux"
 	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
 
 	"go.opentelemetry.io/otel"
 	"go.opentelemetry.io/otel/propagation"
-	sdktrace "go.opentelemetry.io/otel/sdk/trace"
-	"go.opentelemetry.io/otel/sdk/trace/tracetest"
 	"go.opentelemetry.io/otel/trace"
 )
 
@@ -165,48 +161,4 @@ func TestResponseWriterInterfaces(t *testing.T) {
 	}
 
 	router.ServeHTTP(w, r)
-}
-
-func TestCustomSpanNameFormatter(t *testing.T) {
-	exporter := tracetest.NewInMemoryExporter()
-
-	tp := sdktrace.NewTracerProvider(sdktrace.WithSyncer(exporter))
-
-	routeTpl := "/user/{id}"
-
-	testdata := []struct {
-		spanNameFormatter func(string, *http.Request) string
-		expected          string
-	}{
-		{nil, routeTpl},
-		{
-			func(string, *http.Request) string { return "custom" },
-			"custom",
-		},
-		{
-			func(name string, r *http.Request) string {
-				return fmt.Sprintf("%s %s", r.Method, name)
-			},
-			"GET " + routeTpl,
-		},
-	}
-
-	for i, d := range testdata {
-		t.Run(fmt.Sprintf("%d_%s", i, d.expected), func(t *testing.T) {
-			router := mux.NewRouter()
-			router.Use(Middleware("foobar", WithTracerProvider(tp), WithSpanNameFormatter(d.spanNameFormatter)))
-			router.HandleFunc(routeTpl, func(w http.ResponseWriter, r *http.Request) {})
-
-			r := httptest.NewRequest("GET", "/user/123", nil)
-			w := httptest.NewRecorder()
-
-			router.ServeHTTP(w, r)
-
-			spans := exporter.GetSpans()
-			require.Len(t, spans, 1)
-			assert.Equal(t, d.expected, spans[0].Name)
-
-			exporter.Reset()
-		})
-	}
 }
